requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1



cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1

distlib==0.3.7
Django==4.2.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8

django-import-export==3.2.0
django-jazzmin==2.6.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.12.2
html5lib==1.1
idna==3.4
lxml==4.9.3

odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.21.12
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0

pypdf==3.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1

reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.6.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
virtualenv==20.24.2
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
